requirements.txt
absl-py==1.3.0
aiohttp==3.8.3
aiosignal==1.3.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.2.0
build==0.9.0
cachetools==5.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
contourpy==1.0.6
cycler==0.11.0
dill==0.3.6
docopt==0.6.2
etils==0.9.0
flatbuffers==22.12.6
fonttools==4.38.0
frozenlist==1.3.3
gast==0.4.0
google-auth==2.15.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.57.0
grpcio==1.51.1
h5py==3.7.0
idna==3.4
importlib-resources==5.10.1
install==1.3.5
keras==2.11.0
keras-cv==0.3.4
kiwisolver==1.4.4
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
multidict==6.0.4
numpy==1.23.5
oauthlib==3.2.2
openai==0.26.0
opt-einsum==3.3.0
packaging==22.0
pep517==0.13.0
Pillow==9.3.0
pip-tools==6.11.0
pipreqs==0.4.11
promise==2.3
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dateutil==2.8.2
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-addons
tensorflow-datasets==4.7.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.28.0
tensorflow-macos==2.11.0
tensorflow-metadata==1.12.0
tensorflow-metal==0.7.0
termcolor==2.1.1
toml==0.10.2
tomli==2.0.1
tqdm==4.64.1
typing_extensions==4.4.0
urllib3==1.26.13
Werkzeug==2.2.2
wrapt==1.14.1
yarg==0.1.9
yarl==1.8.2
zipp==3.11.0
